+++ src/frontend/apps/lti_site/components/LiveInfoBar/index.spec.tsx
@@ -1,11 +1,31 @@
-import { getDefaultNormalizer, screen } from '@testing-library/react';
+import { getDefaultNormalizer, screen, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import fetchMock from 'fetch-mock';
 import React from 'react';
 import { useParticipantsStore } from 'data/stores/useParticipantsStore/index';
 import render from 'utils/tests/render';
+import { videoMockFactory, report, useJwt } from 'lib-components';
 
 import { LiveInfoBar } from '.';
+import { wrapInVideo } from '../../utils/tests/wrapInVideo';
+
+jest.mock('lib-components', () => ({
+  ...jest.requireActual('lib-components'),
+  report: jest.fn(),
+}));
 
 describe('<LiveInfoBar />', () => {
+  beforeEach(() => {
+    useJwt.setState({
+      jwt: 'json web token',
+    });
+  });
+
+  afterEach(() => {
+    fetchMock.reset();
+    jest.resetAllMocks();
+  });
+
   it('renders placeholder title, startDate and viewers connected', () => {
     useParticipantsStore.setState({
       participants: [
@@ -36,9 +56,19 @@ describe('<LiveInfoBar />', () => {
       ],
     });
 
-    render(<LiveInfoBar title={'title'} startDate={'2022-09-26T07:00:00Z'} />);
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
+    });
 
-    screen.getByRole('heading', { name: 'title' });
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'2022-09-26T07:00:00Z'} />,
+        mockedVideo,
+      ),
+    );
+
+    screen.getByDisplayValue('title');
     screen.getByText('9/26/2022  ·  7:00:00 AM', {
       normalizer: getDefaultNormalizer({ collapseWhitespace: false }),
     });
@@ -46,9 +76,20 @@ describe('<LiveInfoBar />', () => {
   });
 
   it('renders startDate Intl NL - Netherlands', () => {
-    render(<LiveInfoBar title={'title'} startDate={'2022-09-26T07:00:00Z'} />, {
-      intlOptions: { locale: 'nl' },
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
     });
+
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'2022-09-26T07:00:00Z'} />,
+        mockedVideo,
+      ),
+      {
+        intlOptions: { locale: 'nl' },
+      },
+    );
 
     screen.getByText('26-9-2022  ·  07:00:00', {
       normalizer: getDefaultNormalizer({ collapseWhitespace: false }),
@@ -56,9 +97,20 @@ describe('<LiveInfoBar />', () => {
   });
 
   it('renders startDate Intl fr - France', () => {
-    render(<LiveInfoBar title={'title'} startDate={'2022-09-26T07:00:00Z'} />, {
-      intlOptions: { locale: 'fr' },
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
     });
+
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'2022-09-26T07:00:00Z'} />,
+        mockedVideo,
+      ),
+      {
+        intlOptions: { locale: 'fr' },
+      },
+    );
 
     screen.getByText('26/09/2022  ·  07:00:00', {
       normalizer: getDefaultNormalizer({ collapseWhitespace: false }),
@@ -66,10 +118,124 @@ describe('<LiveInfoBar />', () => {
   });
 
   it('renders with invalid startDate', () => {
-    render(<LiveInfoBar title={'title'} startDate={'some date'} />);
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
+    });
 
-    screen.getByRole('heading', { name: 'title' });
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'some date'} />,
+        mockedVideo,
+      ),
+    );
+
+    screen.getByDisplayValue('title');
     expect(screen.queryByText('some date')).not.toBeInTheDocument();
     expect(screen.queryByText('Invalid DateTime')).not.toBeInTheDocument();
+  });
+
+  it('allows you to edit the title', async () => {
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
+    });
+
+    fetchMock.patch(`/api/videos/${mockedVideo.id}/`, {
+      title: ['new title'],
+    });
+
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'some date'} />,
+        mockedVideo,
+      ),
+    );
+
+    const titleInput = screen.getByRole('textbox', {
+      name: 'Enter title of your live here',
+    });
+    titleInput.focus();
+    userEvent.clear(titleInput);
+    userEvent.type(titleInput, 'new title');
+    titleInput.blur();
+    expect(titleInput).toHaveValue('new title');
+    await waitFor(() => expect(fetchMock.calls()).toHaveLength(1));
+    expect(fetchMock.lastCall()![0]).toEqual(`/api/videos/${mockedVideo.id}/`);
+    expect(fetchMock.lastCall()![1]).toEqual({
+      headers: {
+        Authorization: 'Bearer json web token',
+        'Content-Type': 'application/json',
+      },
+      method: 'PATCH',
+      body: JSON.stringify({
+        title: 'new title',
+      }),
+    });
+    expect(titleInput).toHaveValue('new title');
+    screen.getByText('Video updated.');
+  });
+
+  it('should stop user from entering an empty title', async () => {
+    const mockedVideo = videoMockFactory({
+      allow_recording: false,
+      title: 'title',
+    });
+
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'some date'} />,
+        mockedVideo,
+      ),
+    );
+
+    const titleInput = screen.getByRole('textbox', {
+      name: 'Enter title of your live here',
+    });
+    titleInput.focus();
+    userEvent.clear(titleInput);
+    titleInput.blur();
+
+    expect(titleInput).toHaveValue('title');
+    expect(fetchMock.calls()).toHaveLength(0);
+    screen.getByText("Title can't be blank !");
+  });
+
+  it('modifies the input text, but the backend returns an error', async () => {
+    const mockedVideo = videoMockFactory({
+      title: 'An existing title',
+    });
+
+    fetchMock.patch(`/api/videos/${mockedVideo.id}/`, 500);
+
+    render(
+      wrapInVideo(
+        <LiveInfoBar isTeacher startDate={'some date'} />,
+        mockedVideo,
+      ),
+    );
+
+    const textInput = screen.getByRole('textbox', {
+      name: 'Enter title of your live here',
+    });
+    expect(textInput).toHaveValue('An existing title');
+
+    userEvent.type(textInput, ' and more');
+    textInput.blur();
+    await waitFor(() => expect(fetchMock.calls()).toHaveLength(1));
+    expect(fetchMock.lastCall()![0]).toEqual(`/api/videos/${mockedVideo.id}/`);
+    expect(fetchMock.lastCall()![1]).toEqual({
+      headers: {
+        Authorization: 'Bearer json web token',
+        'Content-Type': 'application/json',
+      },
+      method: 'PATCH',
+      body: JSON.stringify({
+        title: 'An existing title and more',
+      }),
+    });
+    expect(textInput).toHaveValue('An existing title');
+    expect(report).toHaveBeenCalled();
+    screen.getByText('Video update has failed !');
   });
 });